+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace `\Heptacom\HeptaConnect\Core\Emission\Contract\EmissionActorInterface` partially with stack processing service described in `\Heptacom\HeptaConnect\Core\Emission\Contract\EmitterStackProcessorInterface`, that is implemented by `\Heptacom\HeptaConnect\Core\Emission\EmitterStackProcessor`
 - Replace `\Heptacom\HeptaConnect\Core\Emission\Contract\EmissionActorInterface` partially with job dispatching from emitter stacks in `\Heptacom\HeptaConnect\Core\Emission\ReceiveJobDispatchingEmitter`
 - Replace `\Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorationActorInterface` partially with stack processing service described in `\Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorerStackProcessorInterface`, that is implemented by `\Heptacom\HeptaConnect\Core\Exploration\ExplorerStackProcessor`
+- Replace `\Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorationActorInterface` partially with job dispatching from explorer stacks in `\Heptacom\HeptaConnect\Core\Exploration\EmissionJobDispatchingExplorer`
 
 ### Deprecated